requirements.txt
pillow==11.1.0
mysql-connector-python==9.2.0